+++ sheraf/attributes/collections.py
@@ -162,6 +162,8 @@ class ListAttribute(sheraf.attributes.Attribute):
         self.persistent_type = persistent_type
         self.accessor_type = accessor_type
         kwargs.setdefault("default", self.persistent_type)
+        kwargs["read_memoization"] = False
+        kwargs["write_memoization"] = False
         super().__init__(**kwargs)
 
     def values(self, list_):